requirements.txt
pandas
requests
sqlalchemy
python-dotenv
psycopg2
matplotlib

pyarrow
azure-storage-blob